Refactor : Seperate utility functions from table
DicomTable.js 파일로부터
descendingComparator, getComparator, stableSort
함수를 분리하여 Utils.js 파일에 추가합니다.

Epic: #27
Task: #31
